Universal Consent Layer
Hello everyone, please find attached a proposition to add a layer of consent to the XMTP protocol. The goal is this document is to propose a way to make user consent interoperable across XMTP clients (for example, if I block someone on a client, this user is also blocked on other clients). Thank you in advance for reviewing this proposition!